% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_openbooks.R
\name{fbopenbooks_server}
\alias{fbopenbooks_server}
\title{Server openbooks}
\usage{
fbopenbooks_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server openbooks
}
\author{
Omar Benites
}

